FreeBSD ARM compilation fix

+++ stat_times_freebsd.go
@@ -12,8 +12,8 @@ func getTimeSpec(info os.FileInfo) timespec {
 	stat := info.Sys().(*syscall.Stat_t)
 	times := timespec{
 		Mtime: info.ModTime(),
-		Atime: time.Unix(stat.Atimespec.Sec, stat.Atimespec.Nsec),
-		Ctime: time.Unix(stat.Ctimespec.Sec, stat.Ctimespec.Nsec),
+		Atime: time.Unix(int64(stat.Atimespec.Sec), int64(stat.Atimespec.Nsec)),
+		Ctime: time.Unix(int64(stat.Ctimespec.Sec), int64(stat.Ctimespec.Nsec)),
 	}
 	return times
 }
